reformat license controller
