+++ controllers/licenseissuer/cmd/main.go
@@ -98,20 +98,6 @@ func main() {
 	// get options for this Operator
 	options := util.GetOptions()
 
-	// if err = (&controller.CollectorReconciler{
-	// 	Client: mgr.GetClient(),
-	// 	Scheme: mgr.GetScheme(),
-	// }).SetupWithManager(mgr); err != nil {
-	// 	setupLog.Error(err, "unable to create controller", "controller", "Collector")
-	// 	os.Exit(1)
-	// }
-	// if err = (&controller.CloudSyncReconciler{
-	// 	Client: mgr.GetClient(),
-	// 	Scheme: mgr.GetScheme(),
-	// }).SetupWithManager(mgr); err != nil {
-	// 	setupLog.Error(err, "unable to create controller", "controller", "CloudSync")
-	// 	os.Exit(1)
-	// }
 	if err = (&controller.LicenseReconciler{
 		Client: mgr.GetClient(),
 		Scheme: mgr.GetScheme(),
