+++ controllers/licenseissuer/internal/controller/license_controller.go
@@ -18,16 +18,19 @@ package controller
 
 import (
 	"context"
-	"os"
+	"errors"
 
 	"github.com/go-logr/logr"
 	accountv1 "github.com/labring/sealos/controllers/account/api/v1"
+
+	notificationv1 "github.com/labring/sealos/controllers/common/notification/api/v1"
 	issuerv1 "github.com/labring/sealos/controllers/licenseissuer/api/v1"
 	"github.com/labring/sealos/controllers/licenseissuer/internal/controller/util"
-	issuer "github.com/labring/sealos/controllers/licenseissuer/internal/manager"
+	ntf "github.com/labring/sealos/controllers/pkg/notification"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
+
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -42,6 +45,10 @@ type LicenseReconciler struct {
 	needRecharge bool
 	logger       logr.Logger
 	Retries      int
+
+	account   accountv1.Account
+	license   issuerv1.License
+	configMap corev1.ConfigMap
 }
 
 //+kubebuilder:rbac:groups=core,resources=secrets,verbs=get;list;watch;create;update;patch;delete
@@ -65,105 +72,37 @@ type LicenseReconciler struct {
 func (r *LicenseReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
 	r.logger.Info("Enter LicenseReconcile", "namespace:", req.Namespace, "name", req.Name)
 	r.logger.Info("Start to get license-related resource...")
-	var canConnectToExternalNetwork bool
-	var (
-		readOperations  issuer.ReadOperationList
-		writeOperations issuer.WriteOperationList
-	)
-	var (
-		license        issuerv1.License
-		uidSecret      corev1.Secret
-		urlConfig      corev1.ConfigMap
-		licenseHistory corev1.ConfigMap
-		account        accountv1.Account
-	)
-	var (
-		ok      bool
-		payload map[string]interface{}
-	)
 
-	canConnectToExternalNetwork = os.Getenv(string(issuer.NetWorkEnv)) == issuer.TRUE
+	// for notification
+	nq := &ntf.NoticeEventQueue{}
+	nm := &ntf.NotificationManager{}
+	nb := (&ntf.Builder{}).WithLevel(notificationv1.High).
+		WithTitle(util.LicenseNoticeTitle).WithFrom(util.Sealos).
+		WithType(ntf.General)
+	receiver := (&ntf.Receiver{}).SetReceiver(req.Namespace)
 
-	// execute read event
-	(&issuer.ReadEventBuilder{}).WithContext(ctx).WithClient(r.Client).
-		WithTag(types.NamespacedName{Namespace: req.Namespace, Name: string(issuer.LicenseName)}).
-		WithObject(&license).AddToList(&readOperations)
-	(&issuer.ReadEventBuilder{}).WithContext(ctx).WithClient(r.Client).
-		WithTag(types.NamespacedName{Namespace: string(issuer.Namespace), Name: string(issuer.ClusterInfoSecretName)}).
-		WithObject(&uidSecret).AddToList(&readOperations)
-	(&issuer.ReadEventBuilder{}).WithContext(ctx).WithClient(r.Client).
-		WithTag(types.NamespacedName{Namespace: string(issuer.Namespace), Name: string(issuer.URLConfigName)}).
-		WithObject(&urlConfig).AddToList(&readOperations)
-	(&issuer.ReadEventBuilder{}).WithContext(ctx).WithClient(r.Client).
-		WithTag(types.NamespacedName{Namespace: string(issuer.Namespace), Name: string(issuer.LicenseHistory)}).
-		WithObject(&licenseHistory).AddToList(&readOperations)
+	reader := &util.Reader{}
+	// get license
 
-	// execute read event
-	if err := readOperations.Execute(); err != nil {
+	reader.Add(&r.license, types.NamespacedName{Namespace: req.Namespace, Name: util.LicenseName})
+	reader.Add(&r.configMap, types.NamespacedName{Namespace: util.SealosNamespace, Name: util.LicenseHistory})
+
+	if err := reader.Read(ctx, r.Client); err != nil {
 		r.logger.Error(err, "failed to read resources...")
 		return ctrl.Result{}, err
 	}
-	// get account
-	err := r.Client.Get(ctx, types.NamespacedName{Namespace: "sealos-system", Name: license.Spec.UID}, &account)
+
+	// check license is valid or not
+	messgae, err := r.Authorize(ctx)
+	nb.WithMessage(messgae).AddToEventQueue(nq)
+	nm.Load(receiver, nq.Events).Run()
 	if err != nil {
-		r.logger.Error(err, "failed to get account")
-		return ctrl.Result{}, err
+		return ctrl.Result{}, r.Delete(ctx, &r.license)
 	}
 
-	config, err := util.ReadConfigFromConfigMap(string(issuer.URLConfigName), &urlConfig)
-	if err != nil {
-		r.logger.Error(err, "failed to read url config")
-		return ctrl.Result{}, err
-	}
-	// security judgement before write event
-	if issuer.CheckLicenseExists(&licenseHistory, license.Spec.Token) {
-		// pack := issuer.NewNotificationPackage(issuer.LicenseNoticeTitle, issuer.SEALOS, issuer.DuplicateLicenseMessage)
-		// issuer.SubmitNotificationWithUser(ctx, r.Client, req.Namespace, pack)
-		return ctrl.Result{}, r.Client.Delete(ctx, &license)
-	}
+	_ = r.RecordLicense(ctx)
 
-	if canConnectToExternalNetwork {
-		payload, ok = issuer.LicenseCheckOnExternalNetwork(license, uidSecret, config.LicenseMonitorURL, r.logger)
-	} else {
-		payload, ok = issuer.LicenseCheckOnInternalNetwork(license)
-	}
-
-	// pre-check for license
-	if !ok {
-		// pack := issuer.NewNotificationPackage(issuer.LicenseNoticeTitle, issuer.SEALOS, issuer.InvalidLicenseMessage)
-		// issuer.SubmitNotificationWithUser(ctx, r.Client, req.Namespace, pack)
-		r.logger.Info("invalid license")
-		return ctrl.Result{}, r.Client.Delete(ctx, &license)
-	}
-
-	// recharge
-	(&issuer.WriteEventBuilder{}).WithCallback(func() error {
-		if !issuer.ContainsFields(payload, issuer.AmountField) {
-			return nil
-		}
-		err := issuer.RechargeByLicense(ctx, r.Client, r.logger, account, payload)
-		if err != nil {
-			// pack := issuer.NewNotificationPackage(issuer.LicenseNoticeTitle, issuer.SEALOS, issuer.RechargeFailedMessage)
-			// issuer.SubmitNotificationWithUser(ctx, r.Client, req.Namespace, pack)
-			return err
-		}
-		// pack := issuer.NewNotificationPackage(issuer.LicenseNoticeTitle, issuer.SEALOS, issuer.ValidLicenseMessage)
-		// issuer.SubmitNotificationWithUser(ctx, r.Client, req.Namespace, pack)
-		return nil
-	}).AddToList(&writeOperations)
-	// record the license
-	(&issuer.WriteEventBuilder{}).WithCallback(func() error {
-		return issuer.RecordLicense(ctx, r.Client, r.logger, license, licenseHistory)
-	}).AddToList(&writeOperations)
-
-	(&issuer.WriteEventBuilder{}).WithCallback(func() error {
-		return r.Client.Delete(ctx, &license)
-	}).AddToList(&writeOperations)
-
-	_ = writeOperations.Execute()
-
-	r.logger.Info("license has been activated")
-	return ctrl.Result{}, nil
+	return ctrl.Result{}, r.Delete(ctx, &r.license)
 }
 
 // SetupWithManager sets up the controller with the Manager.
@@ -173,21 +112,62 @@ func (r *LicenseReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	r.needRecharge = true
 	Predicate := predicate.Funcs{
 		CreateFunc: func(e event.CreateEvent) bool {
-			return e.Object.GetName() == string(issuer.LicenseName)
+			return e.Object.GetName() == string(util.LicenseName)
 		},
 		UpdateFunc: func(e event.UpdateEvent) bool {
-			return e.ObjectNew.GetName() == string(issuer.LicenseName)
+			return e.ObjectNew.GetName() == string(util.LicenseName)
 		},
 		DeleteFunc: func(e event.DeleteEvent) bool {
 			// Ignore delete events
 			return false
 		},
 		GenericFunc: func(e event.GenericEvent) bool {
-			return e.Object.GetName() == string(issuer.LicenseName)
+			return e.Object.GetName() == string(util.LicenseName)
 		},
 	}
 
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&issuerv1.License{}, builder.WithPredicates(Predicate)).
 		Complete(r)
+}
+
+func (r *LicenseReconciler) CheckLicense(ctx context.Context) (string, map[string]interface{}, bool) {
+	options := util.GetOptions()
+	// Check if the license is already used
+	ok := util.CheckLicenseExists(&r.configMap, r.license.Spec.Token)
+	if ok {
+		return util.DuplicateLicenseMessage, nil, false
+	}
+	// Check if the license is valid
+	if options.GetEnvOptions().NetworkConfiguration == "true" {
+		payload, ok := util.LicenseCheckOnExternalNetwork(ctx, r.Client, r.license)
+		if !ok {
+			return util.InvalidLicenseMessage, nil, false
+		}
+		return "", payload, true
+	}
+	payload, ok := util.LicenseCheckOnInternalNetwork(r.license)
+	if !ok {
+		return util.InvalidLicenseMessage, nil, false
+	}
+	return "", payload, true
+}
+
+func (r *LicenseReconciler) Authorize(ctx context.Context) (string, error) {
+	message, payload, ok := r.CheckLicense(ctx)
+	if !ok {
+		return message, errors.New("invalid license")
+	}
+	// recharge
+	if util.ContainsFields(payload, util.AmountField) {
+		err := util.RechargeByLicense(ctx, r.Client, r.account, payload)
+		if err != nil {
+			return util.RechargeFailedMessage, err
+		}
+	}
+	return util.ValidLicenseMessage, nil
+}
+
+func (r *LicenseReconciler) RecordLicense(ctx context.Context) error {
+	return util.RecordLicense(ctx, r.Client, r.license, r.configMap)
 }
