+++ controllers/licenseissuer/internal/controller/util/const.go
@@ -20,6 +20,8 @@ const (
 	SealosNamespace = "sealos-system"
 	ClusterInfo     = "cluster-info"
 	URLConfig       = "url-config"
+	LicenseHistory  = "license-history"
+	LicenseName     = "license"
 	// add more resource name here
 )
 
@@ -56,3 +58,28 @@ const (
 const NoticeFrom = "Sealos Cloud"
 
 const BaseCount = 1000000
+
+const Key = "LS0tLS1CRUdJTiBQVUJMSUMgS0VZLS0tLS0KTUlJQklqQU5CZ2txaGtpRzl3MEJBUUVGQUFPQ0FROEFNSUlCQ2dLQ0FRRUFvbFBTSzB0UjFKeDZtb25lL2ppeApSWGN6UGlxcU5SSXRmdW1mdWNyNGMxc2dqdlJha0NwcWtDU21lMTR1akJkU0x6QlZzRjkvUWl0UnFNb2NvaEN1CkJ6R25EQ29hWnZXbWVHeE96NEZSejVTeUg1QTlDa3dnbUEzYnFnMWxKSEZTMlZyVjVHVFhFWnphZTZtRmhHOVcKenJMTnpZMlptYTMzOVE1WTNJSDZ6RXIrcTRQbTZDOXBHVGpsSnVodlRvb0dSY2w0bmpZRXc2eHB6ZHZrdi9uSApmZmxsWGZVNDNyRGdQaGkwZDRjWnNuTUJlazUxQkNiRFRuSHlNVFdGT1RoTjc1VVM0bzJxRm9JSEhsM0N0RzE4ClZIZEdRSE1IR0dYcGN3bVhhck1EQndwVWFOSk9kMkhjTTB5dlZEY2xDZzRITkIwVUFWeFNweFlRV3BwNWJzN2gKbHdJREFRQUIKLS0tLS1FTkQgUFVCTElDIEtFWS0tLS0tCg=="
+
+const MaxSizeThresholdStr = "800Ki"
+
+const (
+	Sealos                     string = "Sealos Cloud"
+	ClusterCapacityNoticeTitle string = "Attention: Cluster Capacity Alert"
+	LicenseNoticeTitle         string = "Attention: License Issue"
+	InvalidLicenseMessage      string = "The license provided appears to be invalid. Please verify and try again."
+	ExpiredLicenseMessage      string = "The license provided has expired. Please renew and try again."
+	ValidLicenseMessage        string = "Your license has been successfully activated and is now ready for use. Enjoy your Sealos experience!"
+	DuplicateLicenseMessage    string = "The license provided has already been activated. Please use a different license."
+	RechargeFailedMessage      string = "License recharge operation failed."
+)
+
+const (
+	CreatTimeField = "iat"
+	AmountField    = "amt"
+	NodeField      = "nod"
+	CPUField       = "cpu"
+	DurationField  = "tte"
+	AddNodeField   = "and"
+	AddCPUField    = "adc"
+)
